Extract Solana RPC types and fetchSolanaMemoTx into solana.go
